Moved visualization methods into separate module
Methods get attached to class instance
